pyproject.toml
[project]
name = "weather-frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
]

#[tool.uv]
#index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
#trusted-host = "pypi.tuna.tsinghua.edu.cn"
